Add files via upload
In this project I created a Python script to visualize the weather of 900+ random cities across the world,  based on latitude and longitude. To accomplish this, I utilized the Python library citipy, and the OpenWeatherMap API to perform a weather check on each of the cities using a series of successive API calls. Finally, I created scatter plots to easily visualize and showcase the relationships between latitude and temperature (C), Humidity (%), Cloudiness (%), and WInd Speed (mph)